OboeTester: Move setMode inside CommunicationDeviceView
