+++ apps/OboeTester/app/src/main/res/layout/activity_main.xml
@@ -248,19 +248,6 @@
             android:orientation="horizontal">
 
             <TextView
-                android:id="@+id/textView2"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Mode:" />
-
-            <Spinner
-                android:id="@+id/spinnerAudioMode"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:entries="@array/audio_modes"
-                android:prompt="@string/audio_mode_prompt" />
-
-            <TextView
                 android:id="@+id/deviceView"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
